+++ pallets/edge-connect/src/mock.rs
@@ -1,5 +1,4 @@
 pub use crate as pallet_edge_connect;
-use frame_support::traits::ConstU32;
 use frame_support::{derive_impl, weights::constants::RocksDbWeight};
 use frame_system::{mocking::MockBlock, GenesisConfig};
 use pallet_timestamp;
@@ -41,7 +40,6 @@ impl frame_system::Config for Test {
 impl pallet_edge_connect::Config for Test {
 	type RuntimeEvent = RuntimeEvent;
 	type WeightInfo = ();
-	type MaxKycHashLength = ConstU32<64>;
 }
 
 impl pallet_timestamp::Config for Test {
