+++ pallets/edge-connect/src/weights.rs
@@ -38,7 +38,6 @@ pub trait WeightInfo {
 	fn suspend_worker() -> Weight;
     fn ban_worker() -> Weight;
     fn unsuspend_worker() -> Weight;
-	fn submit_kyc_verification() -> Weight;
 }
 
 /// Weights for `pallet_edge_connect` using the Substrate node and recommended hardware.
@@ -104,11 +103,7 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
             .saturating_add(T::DbWeight::get().writes(1_u64))
     }
 
-	fn submit_kyc_verification() -> Weight {
-        Weight::from_parts(10_000, 0)
-            .saturating_add(T::DbWeight::get().reads(1))
-            .saturating_add(T::DbWeight::get().writes(1))
-    }
+
 }
 
 // For backwards compatibility and tests.
@@ -172,10 +167,6 @@ impl WeightInfo for () {
             .saturating_add(RocksDbWeight::get().writes(1_u64))
     }
 
-	fn submit_kyc_verification() -> Weight {
-        Weight::from_parts(10_000, 0)
-            .saturating_add(RocksDbWeight::get().reads(1_u64))
-            .saturating_add(RocksDbWeight::get().writes(1_u64))
-    }
+
 
 }
