+++ pallets/payment/src/mock.rs
@@ -1,5 +1,4 @@
 pub use crate as pallet_payment;
-// use frame_support::traits::ConstU32;
 use frame_support::{derive_impl, parameter_types, weights::constants::RocksDbWeight};
 use frame_system::{mocking::MockBlock, GenesisConfig};
 use pallet_sudo;
@@ -64,12 +63,12 @@ impl pallet_payment::Config for Test {
 	type RuntimeEvent = RuntimeEvent;
 	type Currency = Balances;
 	type WeightInfo = ();
+	type MaxKycHashLength = ConstU32<64>;
 }
 
 impl pallet_edge_connect::Config for Test {
 	type RuntimeEvent = RuntimeEvent;
 	type WeightInfo = ();
-	type MaxKycHashLength = ConstU32<64>;
 }
 
 // Parameter types for the Balances pallet (defines token properties such as existential deposit)
