+++ pallets/payment/src/weights.rs
@@ -46,6 +46,8 @@ pub trait WeightInfo {
 	fn record_usage() -> Weight;
 	fn reward_miner() -> Weight;
 	fn distribute_rewards() -> Weight;
+	fn submit_kyc_verification() -> Weight;
+
 }
 
 /// Weights for `pallet_payment` using the Substrate node and recommended hardware.
@@ -136,6 +138,11 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
             .saturating_add(T::DbWeight::get().writes(1_u64))
     }
 	
+	fn submit_kyc_verification() -> Weight {
+        Weight::from_parts(10_000, 0)
+            .saturating_add(T::DbWeight::get().reads(1))
+            .saturating_add(T::DbWeight::get().writes(1))
+    }
 	
 
 }
@@ -227,4 +234,9 @@ impl WeightInfo for () {
             .saturating_add(RocksDbWeight::get().writes(1_u64))
     }
 	
+	fn submit_kyc_verification() -> Weight {
+        Weight::from_parts(10_000, 0)
+            .saturating_add(RocksDbWeight::get().reads(1_u64))
+            .saturating_add(RocksDbWeight::get().writes(1_u64))
+    }
 }
