+++ pallets/task-management/src/mock.rs
@@ -55,13 +55,13 @@ impl pallet_task_management::Config for Test {
 impl pallet_edge_connect::Config for Test {
 	type RuntimeEvent = RuntimeEvent;
 	type WeightInfo = ();
-	type MaxKycHashLength = ConstU32<64>;
 }
 
 impl pallet_payment::Config for Test {
 	type RuntimeEvent = RuntimeEvent;
 	type Currency = ();
 	type WeightInfo = ();
+	type MaxKycHashLength = ConstU32<64>;
 }
 
 impl pallet_timestamp::Config for Test {
