+++ runtime/src/weights/pallet_payment.rs
@@ -41,6 +41,8 @@ pub trait WeightInfo {
 	fn record_usage() -> Weight;
 	fn reward_miner() -> Weight;
 	fn distribute_rewards() -> Weight;
+	fn submit_kyc_verification() -> Weight;
+
 }
 
 /// Weights for `pallet_payment` using the Substrate node and recommended hardware.
@@ -129,6 +131,12 @@ impl<T: frame_system::Config> pallet_payment::WeightInfo for SubstrateWeight<T> 
         Weight::from_parts(2_000_000_000, 0)
             .saturating_add(T::DbWeight::get().writes(1_u64))
     }
+
+	fn submit_kyc_verification() -> Weight {
+        Weight::from_parts(10_000, 0)
+            .saturating_add(T::DbWeight::get().reads(1))
+            .saturating_add(T::DbWeight::get().writes(1))
+    }
 }
 
 // For backwards compatibility and tests.
@@ -194,4 +202,10 @@ impl WeightInfo for () {
 		Weight::from_parts(16_000_000_000, 0)
 			.saturating_add(RocksDbWeight::get().reads_writes(10, 10))
 	}
+
+	fn submit_kyc_verification() -> Weight {
+        Weight::from_parts(10_000, 0)
+            .saturating_add(RocksDbWeight::get().reads(1_u64))
+            .saturating_add(RocksDbWeight::get().writes(1_u64))
+    }
 }
